backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*",
    "ignore::DeprecationWarning:argon2.*"
]
